# specs/003-line-login-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LINE Login 會員認證系統 API
  description: |
    YouTube Loop Player 的 LINE Login 會員認證與資料管理 API

    ## 認證機制
    - 使用 HTTP-only cookies 儲存 access token
    - 每個請求自動攜帶 cookie,後端驗證 token 有效性
    - Token 過期時返回 401,前端應引導使用者重新登入

    ## 錯誤處理
    所有錯誤回應遵循統一格式:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "使用者友善的錯誤訊息"
      }
    }
    ```

  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api
    description: 開發環境
  - url: https://yourdomain.com/api
    description: 生產環境
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: 更新 Access Token
      description: |
        使用 refresh token 更新 access token
        - 僅在 token 即將過期時使用
        - 實作 token rotation (輪換 refresh token)
      operationId: refreshToken
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Token 更新成功
          headers:
            Set-Cookie:
              schema:
                type: string
                example: access_token=...; HttpOnly; Secure; SameSite=Strict; Max-Age=2592000
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Token 已更新
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /guest-data/migrate:
    post:
      tags:
        - GuestData
      summary: 遷移訪客資料
      description: |
        將訪客期間的播放歷史遷移到會員帳號
        - 僅在會員首次登入時呼叫
        - 接收前端 LocalStorage 的 guest_history 資料
        - 批次寫入 video_library 表
      operationId: migrateGuestData
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - historyData
              properties:
                historyData:
                  type: array
                  description: 訪客播放歷史
                  items:
                    type: object
                    properties:
                      videoId:
                        type: string
                        example: dQw4w9WgXcQ
                      title:
                        type: string
                        example: Rick Astley - Never Gonna Give You Up
                      thumbnail:
                        type: string
                        example: https://i.ytimg.com/vi/dQw4w9WgXcQ/mqdefault.jpg
                      playedAt:
                        type: string
                        format: date-time
                        example: 2025-11-01T10:30:00Z
      responses:
        '200':
          description: 資料遷移成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      migratedCount:
                        type: integer
                        example: 5
                        description: 成功遷移的影片數量
                      skippedCount:
                        type: integer
                        example: 2
                        description: 跳過的影片數量 (已存在於影片庫)
        '400':
          description: 無效的資料格式
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: HTTP-only cookie 儲存的 access token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            message:
              type: string
              example: 錯誤訊息描述
  responses:
    Unauthorized:
      description: 未授權 - Token 無效或已過期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: 請先登入
...